+++ src/pgrn-variables.c
@@ -1,7 +1,7 @@
 #include "pgroonga.h"
 
 #include "pgrn-compatible.h"
-#include "pgrn-customscan.h"
+#include "pgrn-custom-scan.h"
 #include "pgrn-global.h"
 #include "pgrn-groonga.h"
 #include "pgrn-log-level.h"
